+++ dmslogo/utils.py
@@ -10,7 +10,150 @@ Utility functions for plotting.
 import matplotlib.pyplot as plt
 import matplotlib.ticker
 
-import numpy as np
+import numpy
+
+
+class AxLimSetter:
+    """Object to determine axis limits from data.
+
+    Args:
+        datalim_pad (float >= 0)
+            Padding of data limits.
+        include_zero (bool)
+            Extend upper and lower data limits to ensure they include zero?
+        max_from_quantile (`None` or 2-tuple `(quantile, frac)`)
+            Used to set limits based on quantiles of data as described below.
+        min_from_quantile (`None` or 2-tuple `(quantile, frac)`)
+            Used to set limits based on quantiles of data as described below.
+        all_equal_data ('raise' or 2-tuple `(minus_min, plus_max)`)
+            If all the data are equal, it will not be possible to set
+            limits using algorithm below. In this case, raise an error
+            or add the indicated amounts to the data limits.
+
+    The axis limits may just be simple padding of the data limits, but
+    the `max_from_quantile` and `min_from_quantile` arguments allow the
+    option to set limits in a meaningful way so data that are just "noise"
+    of very similar values are plotted far inside the axis limits.
+    Specifically, axis limits are determined by :meth:`AxLimSetter.get_lims`
+    as follows:
+
+      1. The user passes the data, and the upper and lower data limits are
+         determined as the simple min and max of the data.
+
+      2. If `include_zero`, these data limits are adjusted to ensure the lower
+         data limit is <= 0 and the upper data limit is >= 0.
+
+      3. If `max_from_quantile` is not `None`, the upper axis limit is
+         set so that the indicated quantile of the data is `frac` of the
+         way from the lower data limit to the upper axis limit. However,
+         if the data max exceeds the upper axis limit this way, then
+         the upper axis limit is set to the data max.
+
+      4. If `min_from_quantile` is not `None`, similar setting of the
+         lower axis limit is done.
+
+      5. The limits from above are padded by `datalim_pad` to the total
+         extent of the axis on both sides.
+
+    Example that just adds simple padding to data:
+
+    >>> data = [0.5, 0.6, 0.4, 0.4, 0.3, 0.5]
+    >>> setter_default = AxLimSetter()
+    >>> setter_default.get_lims(data)
+    (-0.03, 0.63)
+
+    Now use the `max_from_quantile` option to set an upper limit
+    that substantially exceeds the "noise" of the all-similar values:
+
+    >>> setter_max_quantile = AxLimSetter(max_from_quantile=(0.5, 0.05))
+    >>> setter_max_quantile.get_lims(data)
+    (-0.45, 9.45)
+
+    """
+
+    def __init__(self,
+                 *,
+                 datalim_pad=0.05,
+                 include_zero=True,
+                 max_from_quantile=None,
+                 min_from_quantile=None,
+                 all_equal_data=(-0.001, 0.001),
+                 ):
+        """See main class docstring."""
+        if not isinstance(include_zero, bool):
+            raise ValueError(f"`include_zero` not bool: {include_zero}")
+        self.include_zero = include_zero
+
+        if datalim_pad < 0:
+            raise ValueError(f"`datalim_pad` must be > 0: {datalim_pad}")
+        self._datalim_pad = datalim_pad
+
+        self._quantile_lims = {}
+        for arg, lim in [(max_from_quantile, 'max'),
+                         (min_from_quantile, 'min')]:
+            if arg is None:
+                self._quantile_lims[f"{lim}_from_quantile"] = False
+            elif isinstance(arg, (tuple, list)) and len(arg) == 2:
+                self._quantile_lims[f"{lim}_from_quantile"] = True
+                quantile, frac = arg
+                if not (0 < quantile < 1):
+                    raise ValueError(f"quantile for `{lim}_from_quantile` "
+                                     'must be > 0 and < 1')
+                if not (0 < frac < 1):
+                    raise ValueError(f"frac for `{lim}_from_quantile` "
+                                     'must be > 0 and < 1')
+                self._quantile_lims[f"{lim}_quantile"] = quantile
+                self._quantile_lims[f"{lim}_frac"] = frac
+            else:
+                raise ValueError(f"invalid `{lim}_from_quantile` of {arg}")
+
+        if all_equal_data == 'raise':
+            self._all_equal_data = all_equal_data
+        elif (isinstance(all_equal_data, (list, tuple)) and
+              len(all_equal_data) == 2):
+            if all_equal_data[0] > 0:
+                raise ValueError('first element of `all_equal_data` > 0')
+            if all_equal_data[1] < 0:
+                raise ValueError('second element of `all_equal_data` < 0')
+            if all_equal_data[0] == all_equal_data[1]:
+                raise ValueError('`all_equal_data` cannot be all 0')
+            self._all_equal_data = tuple(all_equal_data)
+        else:
+            raise ValueError(f"invalid `all_equal_data`: {all_equal_data}")
+
+    def get_lims(self, data):
+        """Get 2-tuple `(ax_min, ax_max)` given list or array of data."""
+        datamin = min(data)
+        datamax = max(data)
+
+        if self.include_zero:
+            datamax = max(0, datamax)
+            datamin = min(0, datamin)
+
+        qlims = {'max': datamax, 'min': datamin}
+        for lim, other in [('max', datamin), ('min', datamax)]:
+            if self._quantile_lims[f"{lim}_from_quantile"]:
+                quantile = self._quantile_lims[f"{lim}_quantile"]
+                quantile_val = numpy.quantile(data, quantile)
+                frac = self._quantile_lims[f"{lim}_frac"]
+                qlims[lim] = other + (quantile_val - other) / frac
+        datamax = max(datamax, qlims['max'])
+        datamin = min(datamin, qlims['min'])
+
+        assert datamax >= datamin
+        if datamax == datamin:
+            if self._all_equal_data == 'raise':
+                raise ValueError('data min & max equal, see `all_equal_data`')
+            else:
+                datamin += self._all_equal_data[0]
+                datamax += self._all_equal_data[1]
+
+        extent = datamax - datamin
+        assert extent > 0
+        datamin -= self._datalim_pad * extent
+        datamax += self._datalim_pad * extent
+
+        return (datamin, datamax)
 
 
 def breaksAndLabels(xi, x, n):
@@ -45,7 +188,7 @@ def breaksAndLabels(xi, x, n):
     """
     if len(xi) != len(x):
         raise ValueError('`xi` and `x` differ in length.')
-    if not all(isinstance(i, (int, np.integer)) for i in xi):
+    if not all(isinstance(i, (int, numpy.integer)) for i in xi):
         raise ValueError('xi not integer values')
     xi = list(xi)
     if sorted(set(xi)) != xi:
@@ -137,10 +280,10 @@ def despine(fig=None,
             # clip off the parts of the spines that extend past major ticks
             xticks = ax_i.get_xticks()
             if xticks.size:
-                firsttick = np.compress(xticks >= min(ax_i.get_xlim()),
-                                        xticks)[0]
-                lasttick = np.compress(xticks <= max(ax_i.get_xlim()),
-                                       xticks)[-1]
+                firsttick = numpy.compress(xticks >= min(ax_i.get_xlim()),
+                                           xticks)[0]
+                lasttick = numpy.compress(xticks <= max(ax_i.get_xlim()),
+                                          xticks)[-1]
                 ax_i.spines['bottom'].set_bounds(firsttick, lasttick)
                 ax_i.spines['top'].set_bounds(firsttick, lasttick)
                 newticks = xticks.compress(xticks <= lasttick)
@@ -149,10 +292,10 @@ def despine(fig=None,
 
             yticks = ax_i.get_yticks()
             if yticks.size:
-                firsttick = np.compress(yticks >= min(ax_i.get_ylim()),
-                                        yticks)[0]
-                lasttick = np.compress(yticks <= max(ax_i.get_ylim()),
-                                       yticks)[-1]
+                firsttick = numpy.compress(yticks >= min(ax_i.get_ylim()),
+                                           yticks)[0]
+                lasttick = numpy.compress(yticks <= max(ax_i.get_ylim()),
+                                          yticks)[-1]
                 ax_i.spines['left'].set_bounds(firsttick, lasttick)
                 ax_i.spines['right'].set_bounds(firsttick, lasttick)
                 newticks = yticks.compress(yticks <= lasttick)